IOS-3787: Extract container view & config for SwiftUI previews into separate source files
