+++ Tangem/Preview Content/Mocks/ForComponents/CardsInfo/ViewModels/Common/CardInfoPagePreviewConfig.swift
@@ -1,0 +1,15 @@
+//
+//  CardInfoPagePreviewConfig.swift
+//  Tangem
+//
+//  Created by Andrey Fedorov on 20.07.2023.
+//  Copyright © 2023 Tangem AG. All rights reserved.
+//
+
+import Foundation
+
+struct CardInfoPagePreviewConfig: Identifiable {
+    let id = UUID()
+    let initiallySelectedIndex: Int
+    let hasPullToRefresh: Bool
+}
